move onetouch driver, pt 2
