+++ lib/core/device.js
@@ -27,7 +27,7 @@ var builder = require('../objectBuilder')();
 var timeutils = require('../timeutils');
 
 var dexcomDriver = require('../drivers/dexcomDriver');
-var oneTouchMiniDriver = require('../oneTouchMiniDriver');
+var oneTouchMiniDriver = require('../drivers/oneTouchMiniDriver');
 var abbottFreeStyle = require('../drivers/abbottFreeStyle');
 
 var device = {
