+++ lib/drivers/oneTouchMiniDriver.js
@@ -18,8 +18,8 @@
 var _ = require('lodash');
 var async = require('async');
 var moment = require('moment-timezone');
-var crcCalculator = require('./crc.js');
-var struct = require('./struct.js')();
+var crcCalculator = require('../crc.js');
+var struct = require('../struct.js')();
 
 module.exports = function (config) {
   var cfg = _.clone(config);
